tag and user collection calling weaviate functions refactored with functions manipulating with weaviate in weaviate_search exception handling changed - catching real weaviate exceptions, raising custom exceptions some functions (tag_chunks_with_llm, fetch_chunks_by_collection, ...) optimized
