+++ semant_demo_backend/semant_demo/weaviate_exceptions.py
@@ -3,7 +3,7 @@ class WeaviateError(Exception):
     pass
 
 # Specific exceptions
-class WeaviateConnectionError(WeaviateError):
+class WeaviateConnectError(WeaviateError):
     """
     Raised when unable to connect to Weaviate instance.
     """
